Try platforms

+++ native/lvips/src/libvips.rs
@@ -99,7 +99,7 @@ impl VipsImage {
         unsafe {
             let image = bindings::vips_image_new_from_buffer(
                 buffer.as_ptr() as *const c_void,
-                buffer.len() as u64,
+                buffer.len() as usize,
                 options.as_ptr(),
                 utils::NULL
             );
@@ -285,7 +285,7 @@ impl VipsImage {
     }
 
     pub fn jpeg_buffer( &self ) -> Result<Vec<u8>, String> {
-        let mut buffer_buf_size: u64 = 0;
+        let mut buffer_buf_size: usize = 0;
         let mut buffer_out = null();
 
         unsafe {
@@ -302,7 +302,7 @@ impl VipsImage {
     }
 
     pub fn jpeg_buffer_opts( &self, options: &JpegSaveOptions ) -> Result<Vec<u8>, String> {
-        let mut buffer_buf_size: u64 = 0;
+        let mut buffer_buf_size: usize = 0;
         let mut buffer_out = null();
         let profile = c_string(&options.profile).unwrap();
         let params = globals::get_params().unwrap();
@@ -335,7 +335,7 @@ impl VipsImage {
     }
 
     pub fn png_buffer( &self ) -> Result<Vec<u8>, String> {
-        let mut buffer_buf_size: u64 = 0;
+        let mut buffer_buf_size: usize = 0;
         let mut buffer_out = null();
 
         unsafe {
@@ -352,7 +352,7 @@ impl VipsImage {
     }
 
     pub fn png_buffer_opts( &self, options: &PngSaveOptions ) -> Result<Vec<u8>, String> {
-        let mut buffer_buf_size: u64 = 0;
+        let mut buffer_buf_size: usize = 0;
         let mut buffer_out = null();
         let params = globals::get_params().unwrap();
         let profile = c_string(&options.profile).unwrap();
@@ -384,7 +384,7 @@ impl VipsImage {
     }
 
     pub fn webp_buffer( &self ) -> Result<Vec<u8>, String> {
-        let mut buffer_buf_size: u64 = 0;
+        let mut buffer_buf_size: usize = 0;
         let mut buffer_out = null();
 
         unsafe {
@@ -401,7 +401,7 @@ impl VipsImage {
     }
 
     pub fn webp_buffer_opts( &self, options: &WebPSaveOptions ) -> Result<Vec<u8>, String> {
-        let mut buffer_buf_size: u64 = 0;
+        let mut buffer_buf_size: usize = 0;
         let mut buffer_out = null();
         let params = globals::get_params().unwrap();
         let profile = c_string(&options.profile).unwrap();
